+++ CHANGELOG.md
@@ -17,6 +17,7 @@ IMPORTANT NOTE: This version only works on CACTI 1.x++!
 --- develop ---
 * issue#141: Bug Weathermap - Map Style can't change Link Labels to Bits/Sec
 * issue#152: Dont attempt to include directories from Weathermap config files.
+* issue#157: There is not process interlock in WeatherMap
 * issue: More deprecation warnings in image* functions.
 * feature: Complete most of Internaltionalization